+++ CHANGELOG.md
@@ -7,10 +7,13 @@ All notable changes to this library are documented in this file.
 
 ### API Changes
 
+- Refactored PIREP data model to `pyiem.models.pirep`.
+
 ### New Features
 
 - Add `nwsice` color ramp per suggestion.
 - Retain reference to `CGIModel` schema as `environ["_cgimodel_schema"]`.
+- Parse flight level within PIREPs (#1003).
 - Use `akrherz/iem_database:test_data` container for integration tests.
 
 ### Bug Fixes